+++ index.js
@@ -1,3 +1,33 @@
+'use strict';
+
+/** 
+ * Alpha 
+ *
+ * @param	string
+ * @return	bool
+ */
+exports.alpha = function( str ) {
+	var regex = /^[a-zA-Z]+$/;	
+	var status = regex.test( str );
+
+	return status;
+}
+// --------------------------------------------------------------------
+
+/**
+ * Alpha-numeric
+ *
+ * @param	string
+ * @return	bool
+ */
+exports.alpha_numeric = function( str ) {
+	var regex = /^\w+$/;	
+	var status = regex.test( str );
+
+	return status;
+}
+// --------------------------------------------------------------------
+
 /**
  * Greater than
  *
@@ -6,7 +36,7 @@
  * @return	bool
  */
 exports.greater_than = function( str, comparison ) {
-	status = false;
+	var status = false;
 	if ( !isNaN( str ) ) {
 		if ( str > comparison ) {
 			status = true;
@@ -25,9 +55,53 @@ exports.greater_than = function( str, comparison ) {
  * @return	bool
  */
 exports.greater_than_equal_to = function( str, comparison ) {
-	status = false;
+	var status = false;
 	if ( !isNaN( str ) ) {
 		if ( str >= comparison ) {
+			status = true;
+		}
+	}
+
+	return status;
+}
+// --------------------------------------------------------------------
+
+/**
+ * Latitude
+ *
+ * @param	string
+ * @param	int
+ * @return	bool
+ */
+exports.latitude = function( str ) {
+	var status = false;
+	if ( typeof str == 'number' ) {
+		if ( str < -90 || str > 90 ) {
+			status = false;
+		}
+		else {
+			status = true;
+		}
+	}
+
+	return status;
+}
+// --------------------------------------------------------------------
+
+/**
+ * Longitude
+ *
+ * @param	string
+ * @param	int
+ * @return	bool
+ */
+exports.longitude = function( str ) {
+	var status = false;
+	if ( typeof str == 'number' ) {
+		if ( str < -180 || str > 180 ) {
+			status = false;
+		}
+		else {
 			status = true;
 		}
 	}
@@ -44,7 +118,7 @@ exports.greater_than_equal_to = function( str, comparison ) {
  * @return	bool
  */
 exports.less_than = function( str, comparison ) {
-	status = false;
+	var status = false;
 	if ( !isNaN( str ) ) {
 		if ( str < comparison ) {
 			status = true;
@@ -63,12 +137,26 @@ exports.less_than = function( str, comparison ) {
  * @return	bool
  */
 exports.less_than_equal_to = function( str, comparison ) {
-	status = false;
+	var status = false;
 	if ( !isNaN( str ) ) {
 		if ( str <= comparison ) {
 			status = true;
 		}
 	}
+
+	return status;
+}
+// --------------------------------------------------------------------
+
+/**
+ * Numeric
+ *
+ * @param	string
+ * @return	bool
+ */
+exports.numeric = function( str ) {
+	var regex = /^\d+$/;	
+	var status = regex.test( str );
 
 	return status;
 }
@@ -108,6 +196,37 @@ exports.valid_ip = function( str ) {
 // --------------------------------------------------------------------
 
 /**
+ * Validate URL
+ *
+ * @param	string
+ * @return	bool
+ */
+exports.valid_url = function( str ) {
+	var regex = /(http|https):\/\/(\w+:{0,1}\w*)?(\S+)(:[0-9]+)?(\/|\/([\w#!:.?+=&%!\-\/]))?/;
+	var status = regex.test( str );
+
+	return status;
+}
+// --------------------------------------------------------------------
+
+/**
+ * Performs a Regular Expression match test.
+ *
+ * @param	string
+ * @param	string	regex
+ * @return	bool
+ */
+exports.regex_match = function( str, comparison ) {
+	var flags = comparison.replace( /.*\/([gimy]*)$/, '$1' );
+	var pattern = comparison.replace( new RegExp( '^/(.*?)/' + flags + '$' ), '$1' );
+	var regex = new RegExp( pattern, flags );
+	var status = regex.test( str );
+
+	return status;
+}
+// --------------------------------------------------------------------
+
+/**
  * Required
  *
  * @param	string
@@ -132,17 +251,50 @@ exports.required = function( str ) {
  */
 exports.error_message = function( fn, name, string, comparison = '' ) {
 	switch ( fn ) {
-		case 'ip_address':
-			return 'The ' + name + ' field must contain a valid IP. ';
+		case 'alpha':
+			return 'The ' + name + ' field may only contain alphabetical characters. ';
+		break;
+		case 'alpha_numeric':
+			return 'The ' + name + ' field may only contain alpha-numeric characters. ';
 		break;
 		case 'email':
 			return 'The ' + name + ' field must contain a valid email address. ';
 		break;
+		case 'greater_than':
+			return 'The ' + name + ' field must contain a number greater than ' + comparison + '. ';
+		break;
+		case 'greater_than_equal_to':
+			return 'The ' + name + ' field must contain a number greater than or equal to ' + comparison + '. ';
+		break;
+		case 'latitude':
+			return 'The ' + name + ' field must contain latitude values between -90 to 90 degrees inclusive. ';
+		break;
+		case 'longitude':
+			return 'The ' + name + ' field must contain longitude values between -180 to 180 degrees inclusive. ';
+		break;
+		case 'less_than':
+			return 'The ' + name + ' field must contain a number less than ' + comparison + '. ';
+		break;
+		case 'less_than_equal_to':
+			return 'The ' + name + ' field must contain a number less than or equal to ' + comparison + '. ';
+		break;
+		case 'numeric':
+			return 'The ' + name + ' field must contain only numbers. ';
+		break;
+		case 'ip_address':
+			return 'The ' + name + ' field must contain a valid IP. ';
+		break;
+		case 'regex_match':
+			return 'The ' + name + ' field is not in the correct format. ';
+		break;
 		case 'required':
-			return 'The ' + name + ' field is required.';
+			return 'The ' + name + ' field is required. ';
+		break;
+		case 'url':
+			return 'The ' + name + ' field must contain a valid URL. ';
 		break;
 		default:
-			return fn + ' rules unknown.';
+			return 'The ' + fn + ' rules unknown. ';
 		break;
 	}
 }
@@ -155,16 +307,49 @@ exports.error_message = function( fn, name, string, comparison = '' ) {
  * @param	int
  * @return	function
  */
-exports.fn = function( fnc, string, comparison = '' ) {
-	switch ( fnc ) {
-		case 'ip_address':
-			return ( ( !!string ) ? this.valid_ip( string ) : true);
+exports.fn = function( fn, string, comparison = '' ) {
+	switch ( fn ) {
+		case 'alpha':
+			return ( ( !!string ) ? this.alpha( string ) : true);
+		break;
+		case 'alpha_numeric':
+			return ( ( !!string ) ? this.alpha_numeric( string ) : true);
 		break;
 		case 'email':
 			return ( ( !!string ) ? this.valid_email( string ) : true);
 		break;
+		case 'greater_than':
+			return ( ( !!string ) ? this.greater_than( string, comparison ) : true);
+		break;
+		case 'greater_than_equal_to':
+			return ( ( !!string ) ? this.greater_than_equal_to( string, comparison ) : true);
+		break;
+		case 'latitude':
+			return ( ( !!string ) ? this.latitude( string ) : true);
+		break;
+		case 'longitude':
+			return ( ( !!string ) ? this.longitude( string ) : true);
+		break;
+		case 'less_than':
+			return ( ( !!string ) ? this.less_than( string, comparison ) : true);
+		break;
+		case 'less_than_equal_to':
+			return ( ( !!string ) ? this.less_than_equal_to( string, comparison ) : true);
+		break;
+		case 'numeric':
+			return ( ( !!string ) ? this.numeric( string ) : true);
+		break;
+		case 'ip_address':
+			return ( ( !!string ) ? this.valid_ip( string ) : true);
+		break;
+		case 'regex_match':
+			return ( ( !!string ) ? this.regex_match( string, comparison ) : true);
+		break;
 		case 'required':
 			return this.required( string );
+		break;
+		case 'url':
+			return ( ( !!string ) ? this.valid_url( string ) : true);
 		break;
 		default:
 			return false;
@@ -191,16 +376,13 @@ exports.run = function( setup ) {
 		if ( setup ) {
 			if ( setup.length > 0 ) {
 				result['message'] = 'Unfortunately, some error has occured.';
-				
-				var j = 0;
+
 				setup.forEach( function( data ) {
 					if ( !!data.rules ) {
-						var i = 0;
 						data.rules.split( '|' ).forEach( function( rule ) {
 							var regexp = /\(([^)]+)\)/;
 							var matches = regexp.exec( rule );
-
-							if ( matches == null ) {
+							if ( matches == null && rule != null ) {
 								if ( exports.fn( rule, data.value, '' ) == false ) {
 									result['error_lists'].push( 
 										exports.error_message( rule, data.name, data.value, '' )
@@ -209,19 +391,19 @@ exports.run = function( setup ) {
 								}
 							}
 							else {
-								
-								console.log( i + ' Hehe' );
-								var string_with_replace = String( rule.replace( matches[0], '' ) );
-								//string_with_replace = string_with_replace.replace( '(undefined)' );
-								console.log( i + ' Array' );
-								console.log( 'Function name: ' + string_with_replace );
-								console.log( 'Data: ', matches[1].split( ',' ) );
+								var fn_split = String( rule.replace( matches[0], '' ) );
+								var comparison = matches[1].split( ',' );
+								if ( fn_split != null ) {
+									if ( exports.fn( fn_split, data.value, comparison[0] ) == false ) {
+										result['error_lists'].push( 
+											exports.error_message( fn_split, data.name, data.value, comparison )
+										);
+										error_count++;
+									}
+								}
 							}
-
-							i++;
 						} );
 					}
-					j++;
 				} );
 
 				if ( error_count == 0 ) {
@@ -238,6 +420,5 @@ exports.run = function( setup ) {
 		result['message'] = 'Validaton setup is not object';
 	}
 
-	console.log( result );
 	return result;
 }