+++ History.md
@@ -2,13 +2,26 @@
 ==================
 
   * Adding new function valid_email
+  * Adding new function required
   * Adding new function valid_ip
+  * Adding new function alpha
+  * Adding new function alpha_numeric
+  * Adding new function latitude
+  * Adding new function longitude
+  * Adding new function numeric
+  * Adding new function valid_url
+  * Adding new function regex_match
+  * Adding new function required
+  * Adding License
+  * Adding Readme
 
 1.0.1 / 2019-02-12
 ==================
 
   * Adding new function greater_than
   * Adding new function greater_than_equal_to
+  * Adding new function less_than
+  * Adding new function less_than_equal_to
 
 1.0.0 / 2019-02-12
 ==================